requirements.txt
fastapi==0.116.1
matplotlib==3.10.3
mediapipe==0.10.21
numpy==2.3.2
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
Pillow==11.3.0
uvicorn==0.35.0
